requirements.txt
streamlit==1.29.0
pandas==2.0.3
pillow==10.0.0